Quash some warnings
* Minor fixes to LP interface
* Deprecate and remove array literals
  - Array literals are not C++ compatible
  - Add deprecation wrapper for AA_FAR() macro
  - Remove array literals from aa_test.c
* -Wno-unused-function for flex generated code.
* -Wno-deprecated-declarations for test cases, since it's OK to test
  deprecated functions.
* Allow switch fallthroughs.
* Add missing prototype for aa_tf_duqu_lnv().
* Quash some pedantic warnings about semicolons
* Minor other warning issues.

+++ include/wavefront_internal.h
@@ -59,7 +59,7 @@ struct wf_extra {
 
 
 #define WF_FACE(thing) yyextra->face.thing
-#define SET_WF_FACE(thing) WF_FACE(thing) = atol(yytext)-1
+#define SET_WF_FACE(thing) WF_FACE(thing) = (int32_t)atol(yytext)-1
 #define SKIP_WF_FACE(thing) WF_FACE(thing) = AA_RX_WF_OBJ_FACE_NONE
 
 /**
